// src/app/shared/models/user.ts
export class User {
  EParkingId: number;
  CustomerId: number;
  Id: number;
  Username: string;
  PassWord: string;
  FullName: string;
  Sex: boolean;
  Dob: string;
  Mobile: string;
  Email: string;
  Address: string;
  Zalo: string;
  Skype: string;
  Facebook: string;
  AccountNumber: string;
  Bank: string;
  RoleId: number;
  Image: string;
  Actived: boolean;
  TokenKey = '12345678901';
}